feat: refactor error handling, server setup, and testing
- Replace throw errors with return null in services for better error handling
- Update controllers to handle errors, returning specific status and messages
- Create app.ts to refactor server.ts logic, exporting app with server initialization, ApolloServer, startApolloServer, and stopApolloServer for test usage
- Simplify server.ts by importing app.ts
- Set up Jest and Supertest for testing REST routes in server

+++ src/controllers/userController.ts
@@ -37,6 +37,10 @@ export const getUser: RequestHandler = async (req: Request, res: Response) => {
 
     const user = await userService.getUserById(userId);
 
+    if (!user) {
+      return res.status(404).json({ error: "User not found" });
+    }
+
     return res.status(200).json(user);
   } catch (error) {
     console.error("Error fetching user:", error);
@@ -69,6 +73,12 @@ export const updateUser: RequestHandler = async (
 
     if (userIdFromToken !== userIdFromParams) {
       return res.status(403).json({ error: "Access denied." });
+    }
+
+    const user = await userService.getUserById(userIdFromParams);
+
+    if (!user) {
+      return res.status(404).json({ error: "User not found." });
     }
 
     const updateData: Partial<{
@@ -119,6 +129,12 @@ export const deleteUser: RequestHandler = async (
 
     if (userIdFromToken !== userIdFromParams) {
       return res.status(403).json({ error: "Access denied." });
+    }
+
+    const user = await userService.getUserById(userIdFromParams);
+
+    if (!user) {
+      return res.status(404).json({ error: "User not found." });
     }
 
     await userService.deleteUserById(userIdFromParams);
